requirements.txt
tensorflow==2.18.0
numpy>=1.26.0,<2.1.0
pandas>=2.0.3
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.2
absl-py>=1.0.0
astunparse>=1.6.0
flatbuffers>=24.3.25
gast>=0.2.1,!=0.5.0,!=0.5.1,!=0.5.2
google-pasta>=0.1.1
grpcio>=1.24.3,<2.0
h5py>=3.11.0
keras>=3.5.0
libclang>=13.0.0
ml-dtypes>=0.4.0,<0.5.0
opt-einsum>=2.3.2
protobuf>=3.20.3,<6.0.0dev
requests>=2.21.0,<3
setuptools
tensorboard>=2.18,<2.19
termcolor>=1.1.0
typing-extensions>=3.6.6
wrapt>=1.11.0
